communication between the browser window and the extension
setup the content script to listen to the app change event from single-spa and send an event to the background script

setup a background script that listens to messages from the content script, and forwards that onto the panel

setup the panel.js script to listen to scripts from the background script, and send it as a custom event to the panel's window

setup the panel app to listen to the custom events from panel.js and get the appData again when it happens, then rerender.

simple!