Removed developer from manifest because that breaks chrome
Changed the event we're listening to to be "routing-event" since that event is fired every time reroute() is called -- this fixed a bug with the "reset" button for force mounting/unmounting where if the override returned the same boolean as the original mount function, then "app-change" event wasn't fired and the devtools would be in an inconsistent state.

Added a hack for determining if you're in firefox or chrome, and changing the return value based upon that.

Made sure to clean up after myself in adding and removing event listeners in the devtools window.

Fixed a bug where the built extension wasn't working but dev extension was, since "./panel.html" should have be referring to the one in the build folder.